requirements.txt
bcrypt==3.2.2
Django==3.2.8
django-cors-headers==3.13.0
django-environ==0.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ipython==7.28.0
mysql-connector-python==8.0.29
mysqlclient==2.1.0
Pillow==8.4.0